// types/user.ts
// =========================================================================
// ІНТЕРФЕЙСИ ТИПІЗАЦІЇ КОРИСТУВАЧА (USER TYPES)
// =========================================================================

// Основна сутність користувача, яка повертається сервером при успішній авторизації
export interface User {
  email: string;
  username: string;
  avatar: string;
}